py/requirements.txt
numpy
opencv_python
Pillow
opencv_contrib_python